Fixes on ProductDto, repository and service

+++ src/main/java/service/ProductService.java
@@ -1,5 +1,6 @@
 package service;
 
+import java.time.LocalDate;
 import java.util.ArrayList;
 
 import model.dto.ProductDto;
@@ -10,24 +11,43 @@ public class ProductService {
 
 	ProductRepository productRepository = new ProductRepository();
 	
-	public Product getProductById(int productId) {
-		return productRepository.findById(productId);
+	public Product getProductById(Integer productId) throws Exception {
+		
+		if(productId != null) {
+			return productRepository.findById(productId);
+		}
+		
+		throw new Exception("Product ID is null");
 	}
 	
-	public ArrayList<Product> getEveryProducts() {
+	public ArrayList<Product> getProducts() {
 		return productRepository.findAll();
 	}
 	
-	public void createProduct(Product product) {
-		productRepository.insert(null);
+	public void createProduct(ProductDto productDto) {
+		Product product = new Product(); // Mapping de ProductDto vers Product
+		product.setCreatedAt(LocalDate.now());
+		product.setCreatedBy("USER");
+		productRepository.save(product);
 	}
 	
-	public void updateProduct(Product product) {
-		productRepository.update(null);
+	public void updateProduct(Integer productId, Product productDto) throws Exception {
+		if(productId != null && productRepository.findById(productId) != null) {
+			Product product = new Product(); // Mapping de ProductDto vers Product
+			product.setUpdatedAt(LocalDate.now());
+			product.setUpdatedBy("USER");
+			productRepository.update(product);
+		}
+		
+		throw new Exception("Product ID is incorrect or doesn't exist");
 	}
 	
-	public void deleteProduct(int productId) {
-		productRepository.deleteById(productId);
+	public void deleteProduct(Integer productId) throws Exception {
+		if(productId != null) {
+			productRepository.deleteById(productId);
+		}
+		
+		throw new Exception("Product ID is incorrect");
 	}
